Network tessting with networkx
Aim of this work was to understand networkx package and how it could be applied to the project.

A graph of the network has been produced, showing the links between different nodes. Each edge has been given a weighting to represent the time it takes for an agent to travel between nodes.

On top of this, a dataframe has been set up to track the positions of the people within the network. This will be further developed after team discussions.