Creating a high level network for agents to move within
Taking work completed so far with networkx and extending it.

The network creation has been put into functions, and the code is now able to deal with a range of different inputs. This will allow us to change the type of network we want.

The second part has been to input the people and track them when they move about the network. This is ongoing.